refactor(clean up): creating relay.go and doing some general clean up
